requirements.txt
azure-functions
pandas
azure-cosmos